+++ src/tools/voxedit/modules/voxedit-ui/WindowTitles.h
@@ -36,3 +36,4 @@
 #define POPUP_TITLE_RENAME_NODE "###renamenodepopup"
 #define POPUP_TITLE_MODEL_UNREFERENCE "###unreferencedmodelpopup"
 #define POPUP_TITLE_LOAD_PALETTE "###selectpalettepopup"
+#define POPUP_TITLE_UV_EDITOR "###uveditorpopup"