+++ src/tools/voxedit/modules/voxedit-util/modifier/brush/BrushType.h
@@ -8,10 +8,10 @@
 
 namespace voxedit {
 
-enum class BrushType { None, Shape, Plane, Stamp, Line, Path, Paint, Text, Select, Max };
+enum class BrushType { None, Shape, Plane, Stamp, Line, Path, Paint, Text, Select, Texture, Max };
 
-static constexpr const char *BrushTypeStr[] = {"None", "Shape", "Plane", "Stamp", "Line",
-											   "Path", "Paint", "Text",	 "Select"};
+static constexpr const char *BrushTypeStr[] = {"None", "Shape", "Plane", "Stamp",  "Line",
+											   "Path", "Paint", "Text",	 "Select", "Texture"};
 static_assert(lengthof(BrushTypeStr) == (int)BrushType::Max, "BrushTypeStr size mismatch");
 
 } // namespace voxedit